chore: cleanup function
